Refactored no appointment case for reschedule
